+++ src/common/components/elements/LanguageSwitcher/index.tsx
@@ -1,16 +1,11 @@
 'use client'
 
 import UButton from '@/common/components/atoms/UButton'
+import useI18n from '@/common/lib/i18n/clientHooks'
 import { languages } from '@/common/lib/i18n/settings'
 import { Language } from '@/common/lib/i18n/types'
 import { styled } from '@/common/lib/mui/theme'
 import { Stack } from '@mui/material'
-import {
-  useParams,
-  usePathname,
-  useRouter,
-  useSearchParams,
-} from 'next/navigation'
 import { memo } from 'react'
 
 const StyledButton = styled(UButton)(({ theme }) => {
@@ -25,36 +20,10 @@ const StyledButton = styled(UButton)(({ theme }) => {
 })
 
 export const LanguageSwitcher = memo(function LanguageSwitcher() {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const pathname = usePathname()
-  const { lang } = useParams<{
-    lang: Language
-  }>()
+  const { changeLanguage, language } = useI18n()
 
   const handleClick = (lang: Language) => {
     changeLanguage(lang)
-  }
-
-  const changeLanguage = (newLang: string) => {
-    const pathParts = pathname.split('/')
-
-    // 只替換第一個出現的語言代碼
-    if (pathParts[1] === lang) {
-      pathParts[1] = newLang
-    }
-
-    const newPath = pathParts.join('/')
-
-    // 構建新的 search params
-    const newSearchParams = new URLSearchParams(searchParams.toString())
-
-    // 組合新的 URL
-    const newUrl = `${newPath}${
-      newSearchParams.toString() ? '?' + newSearchParams.toString() : ''
-    }`
-
-    router.push(newUrl)
   }
 
   return (
@@ -67,7 +36,7 @@ export const LanguageSwitcher = memo(function LanguageSwitcher() {
               key={l}
               onClick={() => handleClick(l)}
               variant="text"
-              disabled={l === lang}
+              disabled={l === language}
             >
               {/** TODO: i18n 語言 */}
               {l}
