+++ src/middleware.ts
@@ -41,12 +41,13 @@ export function middleware(req: NextRequest) {
   if (!lng) lng = fallbackLang
 
   // Redirect if lng in path is not supported
+  const search = req.nextUrl.search
   if (
     !languages.some((loc) => req.nextUrl.pathname.startsWith(`/${loc}`)) &&
     !req.nextUrl.pathname.startsWith('/_next')
   ) {
     return NextResponse.redirect(
-      new URL(`/${lng}${req.nextUrl.pathname}`, req.url)
+      new URL(`/${lng}${req.nextUrl.pathname}${search}`, req.url)
     )
   }
